backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
APScheduler==3.6.3
asgiref==3.9.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cachetools==4.2.2
cbor2==5.8.0
celery==5.6.2
certifi==2025.7.14
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-ranged-fileresponse==0.1.2
django-webpush==0.3.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
frozenlist==1.8.0
gunicorn==23.0.0
http_ece==1.2.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
mssql-django==1.6
multidict==6.7.1
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
py-ubjson==0.16.1
py-vapid==1.9.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==26.0.0
pysmb==1.2.13
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==13.15
pytz==2025.2
pywebpush==2.3.0
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.4
rpds-py==0.30.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
smbprotocol==1.15.0
sqlparse==0.5.3
tornado==6.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.5.0
v8unpack==1.2.3
vine==5.1.0
wcwidth==0.6.0
yarl==1.23.0
zope.interface==8.2
